Return empty dict when savedialog is canceled (#191)
Closes: #183
Currently when the SaveDialog is canceled or exited without saving then the return is "" -- an empty string. This doesn't have a get method, so it results in a Traceback.

In this PR, I ensure a dict is always returned and in the case of exiting without saving it's just empty.