add lxml_html_clean to testing in setup.cfg (#209)
Tests are failing because pytest appears to import a lot? so it picks up
missing lxml_html_clean from napari

https://github.com/napari/napari-animation/actions/runs/8697301919/job/23852283978#step:7:108

Adding it to `testing` because it's not needed by anything in
napari-animation. Actual use in napari is gated by try:

https://github.com/napari/napari/blob/dec8b09184728b632945ba09c7e6c8fe6c953508/napari/utils/notebook_display.py#L6-L9
But it's needed for napari tests so it's in napari[testing] but the fix
isn't released yet.